Removed star locking upon growing
Made it so that the star no longer locks when it grows. Instead, it will continue to grow and shrink, but the user will only be able to collect the star when its full sized. The button for collecting the star will not work otherwise. Also made it so star speeds up when its time to collect